+++ src/media/peer.rs
@@ -9,13 +9,13 @@ use std::{cell::RefCell, collections::HashMap, convert::TryFrom, fmt, rc::Rc};
 use derive_more::Display;
 use failure::Fail;
 use medea_client_api_proto::{
-    AudioSettings, Direction, MediaType, PeerConnectionState, PeerId as Id,
+    AudioSettings, Direction, IceServer, MediaType, PeerConnectionState, PeerId as Id,
     Track, TrackId, VideoSettings,
 };
 use medea_macro::enum_delegate;
 
 use crate::{
-    api::control::MemberId, media::MediaTrack, signalling::peers::Counter,
+    api::control::MemberId, media::{MediaTrack, IceUser}, signalling::peers::Counter,
 };
 
 /// [`Peer`] doesnt have remote SDP and is waiting for local SDP.
@@ -68,6 +68,9 @@ impl PeerError {
 #[enum_delegate(pub fn partner_peer_id(&self) -> Id)]
 #[enum_delegate(pub fn partner_member_id(&self) -> MemberId)]
 #[enum_delegate(pub fn is_force_relayed(&self) -> bool)]
+#[enum_delegate(pub fn tracks(&self) -> Vec<Track>)]
+#[enum_delegate(pub fn ice_servers_list(&self) -> Option<Vec<IceServer>>)]
+#[enum_delegate(pub fn set_ice_user(&mut self, ice_user: IceUser))]
 #[enum_delegate(pub fn connection_state(&self) -> PeerConnectionState)]
 #[enum_delegate(
     pub fn set_connection_state(
@@ -152,6 +155,7 @@ pub struct Context {
     member_id: MemberId,
     partner_peer: Id,
     partner_member: MemberId,
+    ice_user: Option<IceUser>,
     sdp_offer: Option<String>,
     sdp_answer: Option<String>,
     receivers: HashMap<TrackId, Rc<MediaTrack>>,
@@ -194,7 +198,7 @@ impl<T> Peer<T> {
         self.context.partner_member.clone()
     }
 
-    /// Returns [`Track`]'s of [`Peer`].
+    /// Returns [`Track`]s of this [`Peer`].
     pub fn tracks(&self) -> Vec<Track> {
         let tracks = self.context.senders.iter().fold(
             vec![],
@@ -246,6 +250,16 @@ impl<T> Peer<T> {
     /// Returns [`Peer`] current connection state.
     pub fn connection_state(&self) -> PeerConnectionState {
         *self.context.connection_state.borrow()
+    }
+
+    /// Returns vector of [`IceServer`]s built from this [`Peer`]s [`IceUser`].
+    pub fn ice_servers_list(&self) -> Option<Vec<IceServer>> {
+        self.context.ice_user.as_ref().map(IceUser::servers_list)
+    }
+
+    /// Sets [`IceUser`], which is used to generate [`IceServer`]s
+    pub fn set_ice_user(&mut self, ice_user: IceUser) {
+        self.context.ice_user.replace(ice_user);
     }
 }
 
@@ -331,6 +345,7 @@ impl Peer<Stable> {
             member_id,
             partner_peer,
             partner_member,
+            ice_user: None,
             sdp_offer: None,
             sdp_answer: None,
             receivers: HashMap::new(),