+++ proto/client-api/CHANGELOG.md
@@ -20,8 +20,29 @@ All user visible changes to this project will be documented in this file. This p
 ### Added
 
 - Commands:
-    - `AddPeerConnectionMetrics` with `IceConnectionState` and `PeerConnectionState` metrics ([#71], [#87]);
+    - `AddPeerConnectionMetrics` with `IceConnectionState`, `PeerConnectionState` and `StatsUpdate` metrics ([#71], [#87], [#90]);
     - `UpdateTracks` ([#58]);
+    - `RTCPeerConnection` stats ([#90]):
+        - `RtcCodecStats`;
+        - `RtcInboundRtpStreamStats`;
+        - `RtcOutboundRtpStreamStats`;
+        - `RemoteInboundRtpStreamStat`;
+        - `RemoteOutboundRtpStreamStat`;
+        - `MediaSourceStat`;
+        - `RtpContributingSourceStat`;
+        - `RtcPeerConnectionStat`;
+        - `DataChannelStat`;
+        - `MediaStreamStat`;
+        - `TrackStat`;
+        - `RtcRtpTransceiverStats`;
+        - `SenderStatsKind`;
+        - `ReceiverStatsKind`;
+        - `RtcTransportStats`;
+        - `RtcSctpTransportStats`;
+        - `RtcIceCandidatePairStats`;
+        - `RtcIceCandidateStats`;
+        - `RtcCertificateStats`;
+        - `RtcIceServerStats`.
 - Events: 
     - `RpcSettings` ([#75]);
     - `force_relay` field to `PeerCreated` event ([#79]);
@@ -32,7 +53,6 @@ All user visible changes to this project will be documented in this file. This p
     - `TrackId` and `PeerId` types ([#28]);
     - `Incrementable` trait ([#28]);
     - `CloseReason` and `CloseDescription` types ([#58]).
-    
 
 [#28]: /../../pull/28
 [#58]: /../../pull/58
@@ -42,6 +62,7 @@ All user visible changes to this project will be documented in this file. This p
 [#81]: /../../pull/81
 [#87]: /../../pull/87
 [#89]: /../../pull/89
+[#90]: /../../pull/90